requirements.txt
# Core dependencies for agent system
google-adk>=0.1.0
google-generativeai>=0.3.0
google-cloud-aiplatform>=1.38.0

# Vector search and embeddings
chromadb>=0.4.0
sentence-transformers>=2.2.0

# MCP support
mcp>=0.1.0

# Session and memory
pydantic>=2.0.0

# Observability
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-instrumentation>=0.42b0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0
numpy>=1.24.0
pandas>=2.0.0

# Web UI
flask>=2.3.0
flask-cors>=4.0.0

# Production WSGI Server
gunicorn>=21.2.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

